update: calculation in the price and tax

+++ dtos/assign_participant_dto.go
@@ -8,7 +8,7 @@ type AssignParticipantsRequest struct {
 type AssignedParticipant struct {
 	ParticipantID string `json:"participantId"`
 	ItemID        string `json:"itemId"`
-	AmountOwed    int    `json:"amountOwed"`
+	AmountOwed    float64    `json:"amountOwed"`
 	IsPaid        bool   `json:"isPaid"`
 }
